requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
cloudscraper==1.2.71
frozenlist==1.7.0
idna==3.10
iniconfig==2.1.0
magic-filter==1.0.12
multidict==6.6.3
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pluggy==1.6.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
requests-toolbelt==1.0.0
six==1.17.0
soupsieve==2.7
typing-inspection==0.4.1
typing-extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
python-dotenv>=1.0.0
PyYAML==6.0.2
ruff==0.5.4
black==24.8.0
